float => double

+++ cusim/cuw2v/pycuw2v.py
@@ -75,12 +75,12 @@ class CuW2V:
 
     # normalize word count
     word_count = np.power(self.word_count, self.opt.count_power,
-                          dtype=np.float32)
+                          dtype=np.float64)
     word_count /= np.sum(self.word_count)
     if self.opt.neg:
       self.obj.build_random_table(word_count, self.opt.random_size)
     else:
-      self.obj.build_huffman_tree(word_count)
+      self.obj.build_huffman_tree(word_count.astype(np.float32))
 
     # random initialize alpha and beta
     np.random.seed(self.opt.seed)
